+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## Major Changes
   * Added Circle Directive (Thanks to Abdellatif Ait boudad)
  
+# 0.8.0 (2016-??-??)
+  
+  * added Polygon Directive
+  * [options] property is added for all directives
+